changed methods a little

+++ NewGameJava/src/entities/Entity.java
@@ -17,7 +17,7 @@ public abstract class Entity {
         initHitBox(x, y, width, height);
     }
 
-    public void initHitBox(float x, float y, float width, float height){
+    public void initHitBox(float x, float y, int width, int height){
         hitBox = new Rectangle2D.Float((int) x, (int) y,width,height);
     }
 
@@ -25,11 +25,6 @@ public abstract class Entity {
         g.setColor(Color.PINK);
         g.drawRect((int)hitBox.x,(int)hitBox.y, (int)hitBox.width, (int)hitBox.height);
     }
-
-  //  public void updateHitBox(){
-   //     hitBox.x = (int) x;
-   //     hitBox.y = (int) y;
-  //  }
 
     public Rectangle2D.Float getHitBox() {
         return hitBox;
